+++ examples/data-sources/morpheus_security_package/data-source.tf
@@ -1,0 +1,3 @@
+data "morpheus_security_package" "tf_example_security_package" {
+  name = "tf_example_security_package"
+}