+++ examples/resources/morpheus_security_package/import.sh
@@ -1,0 +1,1 @@
+terraform import morpheus_security_package.tf_example_security_package 1